refactor: split semantic proof model and suite roots
Narrow the semantic-proof authority surface by breaking report\nmodels into conversation/provider/report families and moving suite\nruntime/support logic behind a smaller public entrypoint.\n\n🤖 Generated with [Claude Code](https://claude.com/claude-code)\n\nCo-Authored-By: Claude <noreply@anthropic.com>
